+++ packages/client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @liteforge/client
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.5.0
+  - @liteforge/query@2.0.0
+
 ## 1.0.0
 
 ### Minor Changes