+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/modal
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.5.0
+
 ## 1.1.0
 
 ### Minor Changes