+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/toast
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.5.0
+
 ## 0.2.0
 
 ### Minor Changes